requirements.txt
numpy>=1.24.3
tensorflow-gpu==2.9.0
matplotlib==3.7.1
dpkt==1.9.8
colorama==0.4.4